refactor(file-storage): cut store.rs & service.rs Henry-Kafura coupling
Two edge-dissolving moves that lower the top two HK hubs without splitting
either type (both are legitimate, already-documented hubs):

1. store.rs returns the audit / file-event outbox row types via the repo
   layer (new `repo::{AuditRow, FileEventRow}` aliases defined on `repo/mod`)
   instead of importing the ORM `entity::*` modules directly. Defining the
   aliases on `repo/mod` (a module store already depends on via `Repos`) — not
   re-exporting them from the repo submodules — keeps the dropped edges from
   reappearing. store.rs HK 5_721_600 -> 4_380_600 (fan_out 16 -> 14).

2. `FileEvent` moves from the persistence repo layer
   (`infra::storage::repo::events_outbox_repo`) to the domain layer
   (`domain::audit`, the transactional-outbox domain-records module alongside
   its sibling `AuditEntry`). It is a plain domain record with no persistence
   concerns; living in the repo layer forced `service`, `store`, and `cleanup`
   to each depend on a repo submodule. All three already import `domain::audit`,
   so each drops its `events_outbox_repo` edge with no new edge added — a
   genuine cross-layer dissolve. store.rs HK -> 3_768_700 (fan_out -> 13);
   service.rs HK 4_035_600 -> 3_512_320 (fan_out 15 -> 14).

Cumulative: store.rs -34%, service.rs -13%. No behaviour change; the domain
services no longer reach into the persistence repo layer for their types.
Full test suite green (171 tests), clippy clean, zero test churn.
